requirements.txt
aiofiles==24.1.0
beautifulsoup4==4.13.4
bs4==0.0.2
greenlet==3.2.3
playwright==1.53.0
pyee==13.0.0
python-dateutil==2.9.0.post0
six==1.17.0
soupsieve==2.7
tqdm==4.67.1
typing_extensions==4.14.1